+++ spec/spec_helper.rb
@@ -32,7 +32,6 @@ end
 $LOAD_PATH.unshift "lib"
 require "hanami/utils"
 require "hanami/validations"
-require "hanami/validations/form"
 
 Hanami::Utils.require!("spec/support")
 Hanami::Utils.require!("spec/shared")