+++ spec/unit/hanami/validations/initialize_spec.rb
@@ -1,19 +1,17 @@
 # frozen_string_literal: true
 
-RSpec.describe Hanami::Validations do
+RSpec.describe Hanami::Validator do
   describe "#initialize" do
-    before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+    subject do
+      Class.new(Hanami::Validator) do
         validations do
           required(:attr) { type?(Integer) }
         end
-      end
+      end.new
+    end
 
-      @nested = Class.new do
-        include Hanami::Validations
-
+    let(:nested) do
+      Class.new(Hanami::Validator) do
         validations do
           required(:foo) { filled? }
           required(:num) { type?(Integer) & eql?(23) }
@@ -22,51 +20,56 @@ RSpec.describe Hanami::Validations do
             required(:baz) { filled? }
           end
         end
-      end
+      end.new
+    end
 
-      @params = Class.new do
+    let(:params) do
+      Class.new do
         def initialize(attributes)
           @attributes = Hash[*attributes]
         end
 
-        def to_h
-          @attributes.to_h
+        def to_hash
+          @attributes
         end
       end
     end
 
+    it "returns a frozen object" do
+      expect(subject).to be_frozen
+    end
+
     it "returns a value for the given attribute" do
-      validator = @validator.new(attr: 23)
-      expect(validator.to_h.fetch(:attr)).to eq 23
+      result = subject.call(attr: 23)
+      expect(result.to_h.fetch(:attr)).to eq(23)
     end
 
     it "returns nil when not set" do
-      validator = @validator.new({})
-      expect(validator.to_h.fetch(:attr, :missing)).to eq :missing
+      validator = subject.call({})
+      expect(validator.to_h.fetch(:attr, :missing)).to eq(:missing)
     end
 
     it "accepts any object that implements #to_h" do
-      params    = @params.new([:attr, 23])
-      validator = @validator.new(params)
+      input  = params.new([:attr, 23])
+      result = subject.call(input)
 
-      expect(validator.to_h.fetch(:attr)).to eq 23
+      expect(result.to_h.fetch(:attr)).to eq(23)
     end
 
     it "accepts zero arguments" do
-      validator = @validator.new
-      expect(validator.to_h).to eq({})
+      result = subject.call({})
+      expect(result.to_h).to eq({})
     end
 
     it "doesn't modify the original attributes" do
-      data       = { attr: "23" }
-      validator  = @validator.new(data)
-      validator.validate
+      data = { attr: "23" }
+      subject.call(data)
 
       expect(data[:attr]).to eq("23")
     end
 
     it "accepts symbols as keys, without coercing and whitelisting" do
-      validator = @nested.new(
+      result = nested.call(
         foo:     "ok",
         num:     23,
         unknown: "no",
@@ -76,10 +79,8 @@ RSpec.describe Hanami::Validations do
         }
       )
 
-      result = validator.validate
-
-      expect(result).to be_success
-      expect(result.output).to eq(
+      expect(result).to be_successful
+      expect(result.to_h).to eq(
         foo:     "ok",
         num:     23,
         unknown: "no",
@@ -88,8 +89,6 @@ RSpec.describe Hanami::Validations do
           wat: "oh"
         }
       )
-
-      expect(validator.to_h).to eq(result.output)
     end
   end
 end