+++ spec/unit/hanami/validations/nested_spec.rb
@@ -1,11 +1,9 @@
 # frozen_string_literal: true
 
-RSpec.describe Hanami::Validations do
+RSpec.describe Hanami::Validator do
   describe "nested validations" do
-    before do
-      @validator = Class.new do
-        include Hanami::Validations
-
+    subject do
+      Class.new(Hanami::Validator) do
         validations do
           required(:number) { filled? }
           required(:code) { filled? & eql?("foo") }
@@ -19,28 +17,27 @@ RSpec.describe Hanami::Validations do
             end
           end
         end
-      end
+      end.new
     end
 
     it "returns successful validation result for valid data" do
-      result = @validator.new(number: 23, code: "foo", customer: { name: "Luca", code: "bar", address: { city: "Rome" } }).validate
+      result = subject.call(number: 23, code: "foo", customer: { name: "Luca", code: "bar", address: { city: "Rome" } })
 
-      expect(result).to be_success
-      expect(result.errors).to be_empty
+      expect(result).to be_successful
     end
 
     it "returns failing validation result for invalid data" do
-      result = @validator.new({}).validate
+      result = subject.call({})
 
-      expect(result).not_to be_success
+      expect(result).not_to be_successful
       expect(result.messages.fetch(:number)).to eq ["is missing"]
       expect(result.messages.fetch(:customer)).to eq ["is missing"]
     end
 
     it "returns different failing validations for keys with the same name" do
-      result = @validator.new(code: "x", customer: { code: "y" }).validate
+      result = subject.call(code: "x", customer: { code: "y" })
 
-      expect(result).not_to be_success
+      expect(result).not_to be_successful
       expect(result.messages.fetch(:code)).to eq ["must be equal to foo"]
       expect(result.messages.fetch(:customer).fetch(:code)).to eq ["must be equal to bar"]
     end
@@ -49,7 +46,7 @@ RSpec.describe Hanami::Validations do
     # See https://github.com/hanami/validations/issues/58
     it "safely serialize to nested Hash" do
       data      = { name: "John Smith", address: { line_one: "10 High Street" } }
-      validator = @validator.new(data)
+      validator = subject.call(data)
 
       expect(validator.to_h).to eq(data)
     end
@@ -58,7 +55,7 @@ RSpec.describe Hanami::Validations do
     # See https://github.com/hanami/validations/issues/58#issuecomment-99144243
     it "safely serialize to Hash" do
       data      = { name: "John Smith", tags: [1, 2] }
-      validator = @validator.new(data)
+      validator = subject.call(data)
 
       expect(validator.to_h).to eq(data)
     end